integration_tests/test-requirements.txt
https://github.com/wazo-platform/wazo-test-helpers/archive/master.zip
docker-compose
pyhamcrest
pytest
requests
kombu